Moves `FeedRefreshViewController` to its own file.
